rework fontpreview as an animation
