+++ coldtype/renderable/__init__.py
@@ -1,3 +1,3 @@
-from coldtype.renderable.renderable import Overlay, Action, RenderPass, renderable, skia_direct, fontpreview, iconset, SkiaPen, ColdtypeCeaseConfigException, runnable
-from coldtype.renderable.animation import animation, aframe, FFMPEGExport
+from coldtype.renderable.renderable import Overlay, Action, RenderPass, renderable, skia_direct, iconset, SkiaPen, ColdtypeCeaseConfigException, runnable
+from coldtype.renderable.animation import animation, aframe, FFMPEGExport, fontpreview
 from coldtype.renderable.ui import ui, UIState
