+++ test/visuals/test_fontpreview.py
@@ -1,7 +1,5 @@
 from coldtype.test import *
 
-@fontpreview("/System/Library/Fonts", r"SFNSText", bg=1)
-def preview(r, m):
-    return [
-        StyledString(str(m.stem), Style(str(m), 72)).pens().align(r).f(0),
-    ]
+@fontpreview(r"Times", bg=1)
+def preview(f, m):
+    return StSt(str(m.stem), m, 72).align(f.a.r).f(0)
